src/requirements.txt
opencv-python== 4.6.0.66
matplotlib
numpy
paddleocr
Pillow==9.5.0
PyYAML>=5.3.1
requests>=2.23.0
scipy>=1.4.1
tqdm>=4.41.0
pandas>=1.1.4
seaborn>=0.11.0
easydict
Cython
h5py
six
tb-nightly
future
yacs
gdown
flake8
yapf
isort==4.3.21
imageio
paddlepaddle
protobuf==3.20.*
paho-mqtt
ultralytics
scikit-learn==0.21.0
filterpy==1.4.5